branch test

+++ test.py
@@ -6,5 +6,5 @@ def add(a: int, b: int) -> int:
 def multiply(a: int, b:int) -> int:
     return a * b
 
-def divide(a: int, b: int) -> int:
+def divide(a: int, b: int) -> float:
     return a / b
